move group save to group service
