+++ service/src/main/kotlin/io/provenance/explorer/service/async/AsyncCachingV2.kt
@@ -311,7 +311,7 @@ class AsyncCachingV2(
         saveIbcChannelData(res, txInfo, txUpdate)
         saveSmartContractData(res, txInfo, txUpdate)
         saveNameData(res, txInfo)
-        saveGroups(res, txInfo, txUpdate)
+        groupService.saveGroups(res, txInfo, txUpdate)
         saveSignaturesTx(res, txInfo, txUpdate)
         return TxUpdatedItems(addrs, markers, txUpdate)
     }
@@ -806,242 +806,6 @@ class AsyncCachingV2(
                     }
                 }
             insertList.forEach { NameRecord.insertOrUpdate(it) }
-        }
-    }
-
-    private fun saveGroups(tx: ServiceOuterClass.GetTxResponse, txInfo: TxData, txUpdate: TxUpdate) = transaction {
-        // get groups, save
-        val msgGroups = tx.tx.body.messagesList.mapNotNull { it.getAssociatedGroups() }
-        val eventGroups = tx.txResponse.logsList
-            .flatMap { it.eventsList }
-            .filter { it.type in GroupEvents.values().map { grp -> grp.event } }
-            .flatMap { e ->
-                getGroupEventByEvent(e.type)!!.let {
-                    e.attributesList
-                        .filter { attr -> attr.key in it.idField }
-                        .map { found -> found.value.scrubQuotes().toLong() }
-                }
-            }
-        (msgGroups + eventGroups).toSet().forEach { id ->
-            groupService.buildGroup(id, txInfo)?.let {
-                txUpdate.apply {
-                    if (tx.txResponse.code == 0) {
-                        this.groupsList.add(it)
-                    }
-                    this.groupJoin.add(groupService.buildTxGroup(id, txInfo))
-                }
-            }
-        }
-
-        // get policies, save
-        val msgPolicies = tx.tx.body.messagesList.mapNotNull { it.getAssociatedGroupPolicies() }
-        val eventPolicies = tx.txResponse.logsList
-            .flatMap { it.eventsList }
-            .filter { it.type in GroupPolicyEvents.values().map { pol -> pol.event } }
-            .flatMap { e ->
-                getGroupPolicyEventByEvent(e.type)!!.let {
-                    e.attributesList
-                        .filter { attr -> attr.key in it.idField }
-                        .map { found -> found.value.scrubQuotes() }
-                }
-            }
-        (msgPolicies + eventPolicies).toSet().forEach { addr ->
-            groupService.buildGroupPolicy(addr, txInfo)
-                ?.also { ProcessQueueRecord.insertIgnore(ProcessQueueType.ACCOUNT, addr) }
-                ?.let { policy ->
-                    val (join, savedPolicy) = groupService.buildTxGroupPolicy(addr, txInfo)
-                    txUpdate.apply {
-                        if (tx.txResponse.code == 0) {
-                            this.groupPolicies
-                                .add(listOf(policy, listOf(join).toArray(TxGroupsPolicyTable.tableName)).toObject())
-                        } else if (savedPolicy) {
-                            this.policyJoinAlt.add(join)
-                        }
-                    }
-                }
-        }
-
-        val govProposalMsgTuples = tx.tx.body.messagesList.mapIndexedNotNull { idx, msg -> msg.getAssociatedGroupProposals(idx) }
-        if (tx.txResponse.code == 0) {
-            govProposalMsgTuples.forEach { triple ->
-                when (triple.second) {
-                    GroupGovMsgType.PROPOSAL -> {
-                        // Have to find the proposalId in the log events
-                        val proposalId = tx.mapEventAttrValues(
-                            triple.first,
-                            GroupProposalEvents.GROUP_SUBMIT_PROPOSAL.event,
-                            GroupProposalEvents.GROUP_SUBMIT_PROPOSAL.idField.toList()
-                        )[GroupProposalEvents.GROUP_SUBMIT_PROPOSAL.idField.first()]!!.toLong()
-
-                        val msg = triple.third.toMsgSubmitProposalGroup()
-                        val nodeData = groupService.proposalAtHeight(proposalId, txInfo.blockHeight)?.proposal
-                        val policy = groupService.policyAtHeight(msg.groupPolicyAddress, txInfo.blockHeight)!!.info
-                        val group = groupService.groupAtHeight(policy.groupId, txInfo.blockHeight)!!.info
-                        val tally = groupService.proposalTallyAtHeight(proposalId, txInfo.blockHeight)?.tally
-                        val data = GroupsProposalData(
-                            msg.proposersList,
-                            msg.metadata,
-                            msg.messagesList,
-                            msg.exec.name,
-                            txInfo.txTimestamp,
-                            group.version,
-                            policy.version,
-                            tally,
-                            nodeData?.votingPeriodEnd?.toDateTime()
-                        )
-                        val status = nodeData?.status ?: ProposalStatus.PROPOSAL_STATUS_ACCEPTED
-                        val execResult = nodeData?.executorResult ?: tx.mapEventAttrValues(
-                            triple.first,
-                            GroupProposalEvents.GROUP_EXEC.event,
-                            listOf("result")
-                        )["result"]!!.getGroupsExecutorResult()
-                        val proposal = groupService.buildProposal(
-                            group.id,
-                            policy.address,
-                            proposalId,
-                            data,
-                            nodeData,
-                            status,
-                            execResult,
-                            txInfo
-                        )
-
-                        val votes = groupService.buildVotes(
-                            group.id,
-                            group.version,
-                            msg.proposersList,
-                            vote {
-                                this.proposalId = proposalId
-                                this.option = Types.VoteOption.VOTE_OPTION_YES
-                                this.metadata = ""
-                            },
-                            txInfo
-                        )
-                        txUpdate.apply {
-                            this.groupProposals.add(proposal)
-                            this.groupVotes.addAll(votes)
-                            this.policyJoinAlt.add(groupService.buildTxGroupPolicy(msg.groupPolicyAddress, txInfo).first)
-                        }
-                    }
-
-                    GroupGovMsgType.VOTE -> {
-                        val msg = triple.third.toMsgVoteGroup()
-                        val proposal = groupService.getProposalById(msg.proposalId)!!
-
-                        groupService.buildVotes(
-                            proposal.groupId.toLong(),
-                            proposal.proposalData.groupVersion,
-                            listOf(msg.voter),
-                            vote {
-                                this.proposalId = msg.proposalId
-                                this.option = msg.option
-                                this.metadata = msg.metadata
-                            },
-                            txInfo
-                        ).let {
-                            txUpdate.apply {
-                                this.groupVotes.addAll(it)
-                                this.policyJoinAlt.add(groupService.buildTxGroupPolicy(proposal.policyAddress, txInfo).first)
-                            }
-                        }
-
-                        val execResult = tx.mapEventAttrValues(
-                            triple.first,
-                            GroupProposalEvents.GROUP_EXEC.event,
-                            listOf("result")
-                        )["result"]?.getGroupsExecutorResult()
-
-                        if (execResult != null) {
-                            transaction {
-                                proposal.apply {
-                                    if (proposal.proposalStatus.getGroupsProposalStatus() != ProposalStatus.PROPOSAL_STATUS_ACCEPTED) {
-                                        this.proposalStatus = ProposalStatus.PROPOSAL_STATUS_ACCEPTED.name
-                                    }
-                                    if (proposal.executorResult.getGroupsExecutorResult() != execResult) {
-                                        this.executorResult = execResult.name
-                                    }
-                                }
-                            }
-                        }
-                    }
-
-                    GroupGovMsgType.EXEC -> {
-                        val msg = triple.third.toMsgExecGroup()
-                        val proposal = groupService.getProposalById(msg.proposalId)!!
-
-                        txUpdate.apply {
-                            this.policyJoinAlt.add(groupService.buildTxGroupPolicy(proposal.policyAddress, txInfo).first)
-                        }
-
-                        val execResult = tx.mapEventAttrValues(
-                            triple.first,
-                            GroupProposalEvents.GROUP_EXEC.event,
-                            listOf("result")
-                        )["result"]?.getGroupsExecutorResult()
-
-                        if (execResult != null) {
-                            transaction {
-                                proposal.apply {
-                                    if (proposal.proposalStatus.getGroupsProposalStatus() != ProposalStatus.PROPOSAL_STATUS_ACCEPTED) {
-                                        this.proposalStatus = ProposalStatus.PROPOSAL_STATUS_ACCEPTED.name
-                                    }
-                                    if (proposal.executorResult.getGroupsExecutorResult() != execResult) {
-                                        this.executorResult = execResult.name
-                                    }
-                                }
-                            }
-                        }
-                    }
-
-                    GroupGovMsgType.WITHDRAW -> {
-                        val msg = triple.third.toMsgWithdrawProposalGroup()
-                        val proposal = groupService.getProposalById(msg.proposalId)!!
-
-                        txUpdate.apply {
-                            this.policyJoinAlt.add(groupService.buildTxGroupPolicy(proposal.policyAddress, txInfo).first)
-                        }
-
-                        transaction {
-                            proposal.apply { this.proposalStatus = ProposalStatus.PROPOSAL_STATUS_WITHDRAWN.name }
-                        }
-                    }
-                }
-            }
-        } else {
-            govProposalMsgTuples.forEachIndexed { _, triple ->
-                when (triple.second) {
-                    GroupGovMsgType.PROPOSAL -> {
-                        val msg = triple.third.toMsgSubmitProposalGroup()
-                        txUpdate.apply {
-                            this.policyJoinAlt.add(groupService.buildTxGroupPolicy(msg.groupPolicyAddress, txInfo).first)
-                        }
-                    }
-
-                    GroupGovMsgType.VOTE -> {
-                        val msg = triple.third.toMsgVoteGroup()
-                        val proposal = groupService.getProposalById(msg.proposalId)!!
-                        txUpdate.apply {
-                            this.policyJoinAlt.add(groupService.buildTxGroupPolicy(proposal.policyAddress, txInfo).first)
-                        }
-                    }
-
-                    GroupGovMsgType.EXEC -> {
-                        val msg = triple.third.toMsgExecGroup()
-                        val proposal = groupService.getProposalById(msg.proposalId)!!
-                        txUpdate.apply {
-                            this.policyJoinAlt.add(groupService.buildTxGroupPolicy(proposal.policyAddress, txInfo).first)
-                        }
-                    }
-
-                    GroupGovMsgType.WITHDRAW -> {
-                        val msg = triple.third.toMsgWithdrawProposalGroup()
-                        val proposal = groupService.getProposalById(msg.proposalId)!!
-                        txUpdate.apply {
-                            this.policyJoinAlt.add(groupService.buildTxGroupPolicy(proposal.policyAddress, txInfo).first)
-                        }
-                    }
-                }
-            }
         }
     }
